% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_alphas_for_scales}
\alias{calc_alphas_for_scales}
\title{Calc Cronbach's alpha for multiple scales}
\usage{
calc_alphas_for_scales(data, scales, values = c(1, 5), conf.level = 0.95)
}
\arguments{
\item{data}{raw data as data.frame object}

\item{scales}{data.frame object describing scales and which items it consists
of. The structure is : \code{data.frame(scale = "item_1, -item_2, ... , item_n")},
where the negatives means that item is need to be reversed}

\item{values}{\code{c(min, max)} value of item}

\item{conf.level}{the confidence level for confidence intervals}
}
\value{
A data.frame containing number of items, raw Cronbachs alphas, standarized
alphas and lower and upper CI for alphas
}
\description{
Calc Cronbach's alpha for multiple scales
}
