% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_items_to_scale}
\alias{calc_items_to_scale}
\title{Calculate sum/or means for multiple scales of questionnaire}
\usage{
calc_items_to_scale(data, scales, values = c(1, 5), mean = TRUE, na.rm = TRUE)
}
\arguments{
\item{data}{raw data as a data.frame object}

\item{scales}{data.frame object describing scales and which items it consists
of. The structure is : \code{data.frame(scale = "item_1, -item_2, ... , item_n")},
where the negatives means that item is need to be reversed}

\item{values}{\code{c(min, max)} value of item}

\item{mean}{should the function calculate means of scales? If \code{FALSE}, then
the sums will be calculated}

\item{na.rm}{remove missing values?}
}
\value{
data.frame containing the summarized scales or mean of items in scales
}
\description{
Calculate sum/or means for multiple scales of questionnaire
}
