% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_scale_desc}
\alias{calc_scale_desc}
\title{Calculate scale descriptives for one or more variables with or without grouping}
\usage{
calc_scale_desc(
  data,
  Vars.list,
  grouping = "no grouping",
  not.normal.list = FALSE,
  normal.list = FALSE,
  na.rm = FALSE,
  normality_test = "shapiro"
)
}
\arguments{
\item{data}{data.frame object}

\item{Vars.list}{character vector of variable names, for which the scale
descriptives will be calculated}

\item{grouping}{name of the grouping variable. defaults to "no grouping", where
there will not be any grouping}

\item{not.normal.list}{if \code{TRUE}, then output will be a character vector
with names of not normal distributed variables}

\item{normal.list}{if \code{TRUE}, then output will be a character vector with
names of normally distributed variable}

\item{na.rm}{remove missing values}

\item{normality_test}{which normality test to use for checking normality of
distribution. Defaults to \code{"shapiro"}, can also use \code{"lillie"} for
Lilliefors test.}
}
\description{
Calculate scale descriptives for one or more variables with or without grouping
}
