% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{export_mass_xlsx}
\alias{export_mass_xlsx}
\title{Export list of dataframes as multiple sheets of xlsx file}
\usage{
export_mass_xlsx(list_of_dfs, xlsx_name)
}
\arguments{
\item{list_of_dfs}{list of dataframes to export}

\item{xlsx_name}{character string specyfying name of exported .xlsx}
}
\description{
Export list of dataframes as multiple sheets of xlsx file
}
