% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexes.R
\name{flex_corr}
\alias{flex_corr}
\title{Calculate correlations and present in flextable. Automatically tests normality to choose between Pearson or Spearman}
\usage{
flex_corr(
  data,
  vars_hor,
  vars_ver,
  vars_hor_names = vars_hor,
  vars_ver_names = vars_ver,
  normality_test = "shapiro",
  adjust = "none",
  lang = "english",
  p_values = "include",
  col_widths = "default"
)
}
\arguments{
\item{vars_hor}{names of columns in data on horizontal}

\item{vars_ver}{names of columns in data on vertical}

\item{vars_hor_names}{names vars_hor to print}

\item{vars_ver_names}{names of vars_ver to print}

\item{normality_test}{which normality test to use? Either "shapiro" or "lillie"}

\item{adjust}{adjustement for multiple tests}

\item{lang}{default to "english". Can also be "polish"}

\item{p_values}{include p values?}

\item{col_widths}{how width should the cols be? If not default, then specify
list with numeric values for \code{r_width}, \code{m_width}, \code{p_width} and \code{s_width}}
}
\value{
flextable
}
\description{
Calculate correlations and present in flextable. Automatically tests normality to choose between Pearson or Spearman
}
