% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexes.R
\name{flex_reg}
\alias{flex_reg}
\title{Calculate multiple linear regression and generate flextable.}
\usage{
flex_reg(
  data,
  ind_var,
  pred_vars,
  ind_name,
  pred_names,
  formula = NULL,
  lang = "english",
  width_adjust = 1,
  fontsize = 9
)
}
\arguments{
\item{data}{data.frame}

\item{ind_var}{name of column containing dependent variable}

\item{pred_vars}{names of columns containing indepented (predictory) variables}

\item{ind_name}{name of dependent variable for printing}

\item{pred_names}{names of independent (predictory) variables for printing}

\item{formula}{optionally formula to use - if there are something more complex to calculate}

\item{lang}{defaults to "english", may also be "polish"}

\item{width_adjust}{for adjusting width of columns}

\item{fontsize}{defaults to 9}
}
\description{
Calculate multiple linear regression and generate flextable.
}
