% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexes.R
\name{flex_wilcox}
\alias{flex_wilcox}
\title{Generate Wilcoxon test of differences between two groups and present results as a flextable}
\usage{
flex_wilcox(
  data,
  vars,
  grouping,
  vars_names = NULL,
  grouping_names = NULL,
  group_name = NULL,
  lang = "english",
  alternative = "two.sided",
  h_lines = FALSE
)
}
\arguments{
\item{vars}{names of columns in data to test for differences}

\item{grouping}{name of column in data to distinguish groups}

\item{vars_names}{names of variables to print in flextable}

\item{grouping_names}{name of groups}

\item{group_name}{name of grouping in header}

\item{lang}{default to "english". Can also be "polish"}

\item{alternative}{either "two.sided" (default), "greater" or "less"}

\item{h_lines}{if there should be horizontal lines between variables}
}
\value{
flextable
}
\description{
Generate Wilcoxon test of differences between two groups and present results as a flextable
}
