% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{summarize_descriptives}
\alias{summarize_descriptives}
\title{Summarize descriptives: mean and standard deviation for grouped variable}
\usage{
summarize_descriptives(data, var, grouping.var)
}
\arguments{
\item{data}{data.frame}

\item{var}{name of the variable for which to calculate descriptives}

\item{grouping.var}{names of the variables which should be used for grouping,
maximum 2}
}
\description{
Summarize descriptives: mean and standard deviation for grouped variable
}
